requirements.txt
flask>=3.0.0
torch>=2.10.0
torchvision>=0.16.0
opencv-python>=4.8.1.78
Pillow>=10.1.0
numpy>=1.24.3
werkzeug>=3.0.1
facenet-pytorch>=2.5.3
tqdm>=4.66.1
albumentations>=1.3.1
matplotlib>=3.8.0
scikit-learn>=1.3.2
pandas>=2.1.3